Remove webpack call from env.sh template
The `env.sh` template creates `env.sh` on the prod server and runs as root. This fires webpack in the `app` folder but creates the `app/public/js/app.js` folder with `root` permissions. 

`env.sh` is sourced during deployment in `web.tpl`, causing the wrong permissions to be applied. Removing these lines from this template solves the issue.